refactor: 자동차 이름 객체(문자열) 포장
