+++ src/main/java/racingcar/domain/Cars.java
@@ -37,7 +37,7 @@ public class Cars {
     public String getNames() {
         List<String> names = new ArrayList<>();
         for (Car car : cars) {
-            names.add(car.getName());
+            names.add(car.getNameToString());
         }
         return String.join(",", names);
     }
