+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,31 +1,23 @@
 package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.Test;
 
 public class CarTest {
 
     @Test
-    void 자동차_이름_최대길이_초과() {
-        assertThatThrownBy(
-                () -> new Car("pikachu"))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
     void 자동차_움직임() {
         Car car = new Car("yoshi");
         car.tryToMove(5);
-        assertThat(car.getPosition()).isEqualTo(1);
+        assertThat(car.getPosition()).isEqualTo(new Position(1));
     }
 
     @Test
     void 자동차_멈춤() {
         Car car = new Car("kirby");
         car.tryToMove(2);
-        assertThat(car.getPosition()).isEqualTo(0);
+        assertThat(car.getPosition()).isEqualTo(new Position());
     }
 
     @Test
